Experimental Rotational and Camera changes
A new class Rotational has conversions between three types of rotations.
I'm not sure if this is right way to go about it (especially the use of
Euler angles), but I've found it useful when working with the sdl
backend.

Cameras are now represented by yaw, pitch and roll, using the OpenGL
coordinate convention. The problem with previous forward vector
representation is that when forward aligns with up, you don't know which
way you're facing. Also there was no roll support (less crucial).
Supporting other coordinate systems would be ideal but for now I'll
stick to this convention and see how it goes.