Change the internals of Direction and fix some envelope functions
This makes is nicer for functions that accept directions because they
longer have to signorm the input whereas there are ways to construct
unit vectors without signorm (like angleDir). This does have the
disadvantage that transformations always need to be signormed but this
is somewhat rare and could be addressed with rewrite rules (or maybe
special functions like rotateDir).

Also some of the envelope functions where broken because the internals
of envelope except a unit vector. These have returned to their old
semantics and Dir versions have been added, at least the ones I caught.
Everything will need proper testing.